implement computeCost function
this was a bit confusing for me because i thought the way to calculate the hypothesis would be like this:

h = theta.' * X;

because the hypothesis funciton was defined in the course as the result of the transpose of theta times X.

however, that line of code resulted in errors from Octave due to a mismatch in the dimensions of the operands for the `*` operation (because the col of op1 should match the row of op2 in matrix multiplicaiton).

anyway

this works... I'll figure out more Octave syntax later. The math makes sense